+++ packages/rrdom-nodejs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # rrdom-nodejs
 
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @saola.ai/rrweb-snapshot@2.0.6
+  - @saola.ai/rrdom@2.0.6
+
 ## 2.0.5
 
 ### Patch Changes